+++ backend/src/main/java/org/donggle/backend/application/service/writing/MemberCategoryNotionInfo.java
@@ -1,4 +1,4 @@
-package org.donggle.backend.application.service;
+package org.donggle.backend.application.service.writing;
 
 import org.donggle.backend.domain.category.Category;
 import org.donggle.backend.domain.member.Member;
